# Java Extension Optimizations