--- a/java/src/json/ext/StringEncoder.java
+++ b/java/src/json/ext/StringEncoder.java
@@ -249,11 +249,11 @@ class StringEncoder extends ByteListTranscoder {
 
         if (pos + 4 <= len) {
             int x = bb.getInt(ptr + pos);
-            int is_ascii = 0x808080 & ~x;
-            int xor2 = x ^ 0x020202;
-            int lt32_or_eq34 = xor2 - 0x212121;
-            int sub92 = x ^ 0x5C5C5C;
-            int eq92 = (sub92 - 0x010101);
+            int is_ascii = 0x80808080 & ~x;
+            int xor2 = x ^ 0x02020202;
+            int lt32_or_eq34 = xor2 - 0x21212121;
+            int sub92 = x ^ 0x5C5C5C5C;
+            int eq92 = (sub92 - 0x01010101);
             boolean skip_chunk = ((lt32_or_eq34 | eq92) & is_ascii) == 0;
             if (skip_chunk) {
                 pos += 4;