## Changelog 📓 
- Use a segmented buffer for the `OutputStream` to reduce `System.arraycopy`'s each time the output buffer is resized.
- Refactored `StringEncoder#encode` to include a [SWAR-based](https://en.wikipedia.org/wiki/SWAR) fast path for basic JSON encoding. The algorithm is from [this post](https://lemire.me/blog/2025/04/13/detect-control-characters-quotes-and-backslashes-efficiently-using-swar/). It's the same as the vector-based algorithm in the C extension. 

These features can be toggled with the system properties `json.useSegmentedOutputStream` and `json.useSWARBasicEncoder`. Both default to `true`. I'm happy to remove these. They made testing and benchmarking much easier. 

## Benchmarks

### SegmentedByteListDirectOutputStream + SWAR
```
% ONLY=json JAVA_OPTS='-Djson.useSegmentedOutputStream=true -Djson.useSWARBasicEncoder=true' ruby -I"lib" benchmark/encoder-realworld.rb

== Encoding activitypub.json (52595 bytes)
jruby 9.4.12.0 (3.1.4) 2025-02-11 f4ab75096a OpenJDK 64-Bit Server VM 24.0.1+9-30 on 24.0.1+9-30 +jit [arm64-darwin]
Warming up --------------------------------------
                json     1.741k i/100ms
Calculating -------------------------------------
                json     18.378k (± 6.3%) i/s   (54.41 μs/i) -    182.805k in  10.011722s

== Encoding citm_catalog.json (500298 bytes)
jruby 9.4.12.0 (3.1.4) 2025-02-11 f4ab75096a OpenJDK 64-Bit Server VM 24.0.1+9-30 on 24.0.1+9-30 +jit [arm64-darwin]
Warming up --------------------------------------
                json    85.000 i/100ms
Calculating -------------------------------------
                json    857.615 (± 1.3%) i/s    (1.17 ms/i) -      8.585k in  10.012075s

== Encoding twitter.json (466906 bytes)
jruby 9.4.12.0 (3.1.4) 2025-02-11 f4ab75096a OpenJDK 64-Bit Server VM 24.0.1+9-30 on 24.0.1+9-30 +jit [arm64-darwin]
Warming up --------------------------------------
                json   185.000 i/100ms
Calculating -------------------------------------
                json      1.849k (± 1.0%) i/s  (540.77 μs/i) -     18.500k in  10.005181s

== Encoding ohai.json (20147 bytes)
jruby 9.4.12.0 (3.1.4) 2025-02-11 f4ab75096a OpenJDK 64-Bit Server VM 24.0.1+9-30 on 24.0.1+9-30 +jit [arm64-darwin]
Warming up --------------------------------------
                json     2.558k i/100ms
Calculating -------------------------------------
                json     25.217k (± 1.1%) i/s   (39.66 μs/i) -    253.242k in  10.043890s
```


### ByteListDirectOutputStream + SWAR

```
% ONLY=json JAVA_OPTS='-Djson.useSegmentedOutputStream=false -Djson.useSWARBasicEncoder=true' ruby -I"lib" benchmark/encoder-realworld.rb

== Encoding activitypub.json (52595 bytes)
jruby 9.4.12.0 (3.1.4) 2025-02-11 f4ab75096a OpenJDK 64-Bit Server VM 24.0.1+9-30 on 24.0.1+9-30 +jit [arm64-darwin]
Warming up --------------------------------------
                json     1.560k i/100ms
Calculating -------------------------------------
                json     15.622k (± 0.8%) i/s   (64.01 μs/i) -    157.560k in  10.086737s

== Encoding citm_catalog.json (500298 bytes)
jruby 9.4.12.0 (3.1.4) 2025-02-11 f4ab75096a OpenJDK 64-Bit Server VM 24.0.1+9-30 on 24.0.1+9-30 +jit [arm64-darwin]
Warming up --------------------------------------
                json    87.000 i/100ms
Calculating -------------------------------------
                json    875.692 (± 0.9%) i/s    (1.14 ms/i) -      8.787k in  10.035282s

== Encoding twitter.json (466906 bytes)
jruby 9.4.12.0 (3.1.4) 2025-02-11 f4ab75096a OpenJDK 64-Bit Server VM 24.0.1+9-30 on 24.0.1+9-30 +jit [arm64-darwin]
Warming up --------------------------------------
                json   182.000 i/100ms
Calculating -------------------------------------
                json      1.818k (± 0.8%) i/s  (550.15 μs/i) -     18.200k in  10.013389s

== Encoding ohai.json (20147 bytes)
jruby 9.4.12.0 (3.1.4) 2025-02-11 f4ab75096a OpenJDK 64-Bit Server VM 24.0.1+9-30 on 24.0.1+9-30 +jit [arm64-darwin]
Warming up --------------------------------------
                json     2.544k i/100ms
Calculating -------------------------------------
                json     25.319k (± 0.9%) i/s   (39.50 μs/i) -    254.400k in  10.048804s
```

### ByteListDirectOutputStream + Scalar

```
% ONLY=json JAVA_OPTS='-Djson.useSegmentedOutputStream=false -Djson.useSWARBasicEncoder=false' ruby -I"lib" benchmark/encoder-realworld.rb

== Encoding activitypub.json (52595 bytes)
jruby 9.4.12.0 (3.1.4) 2025-02-11 f4ab75096a OpenJDK 64-Bit Server VM 24.0.1+9-30 on 24.0.1+9-30 +jit [arm64-darwin]
Warming up --------------------------------------
                json     1.078k i/100ms
Calculating -------------------------------------
                json     10.829k (± 2.5%) i/s   (92.35 μs/i) -    108.878k in  10.062513s

== Encoding citm_catalog.json (500298 bytes)
jruby 9.4.12.0 (3.1.4) 2025-02-11 f4ab75096a OpenJDK 64-Bit Server VM 24.0.1+9-30 on 24.0.1+9-30 +jit [arm64-darwin]
Warming up --------------------------------------
                json    78.000 i/100ms
Calculating -------------------------------------
                json    810.901 (± 2.8%) i/s    (1.23 ms/i) -      8.112k in  10.013134s

== Encoding twitter.json (466906 bytes)
jruby 9.4.12.0 (3.1.4) 2025-02-11 f4ab75096a OpenJDK 64-Bit Server VM 24.0.1+9-30 on 24.0.1+9-30 +jit [arm64-darwin]
Warming up --------------------------------------
                json   128.000 i/100ms
Calculating -------------------------------------
                json      1.269k (± 3.3%) i/s  (788.26 μs/i) -     12.672k in  10.001657s

== Encoding ohai.json (20147 bytes)
jruby 9.4.12.0 (3.1.4) 2025-02-11 f4ab75096a OpenJDK 64-Bit Server VM 24.0.1+9-30 on 24.0.1+9-30 +jit [arm64-darwin]
Warming up --------------------------------------
                json     2.178k i/100ms
Calculating -------------------------------------
                json     21.633k (± 1.0%) i/s   (46.23 μs/i) -    217.800k in  10.068853s
```

### SegmentedByteListDirectOutputStream + Scalar

```
% ONLY=json JAVA_OPTS='-Djson.useSegmentedOutputStream=true -Djson.useSWARBasicEncoder=false' ruby -I"lib" benchmark/encoder-realworld.rb

== Encoding activitypub.json (52595 bytes)
jruby 9.4.12.0 (3.1.4) 2025-02-11 f4ab75096a OpenJDK 64-Bit Server VM 24.0.1+9-30 on 24.0.1+9-30 +jit [arm64-darwin]
Warming up --------------------------------------
                json     1.014k i/100ms
Calculating -------------------------------------
                json     10.203k (± 0.8%) i/s   (98.01 μs/i) -    102.414k in  10.037929s

== Encoding citm_catalog.json (500298 bytes)
jruby 9.4.12.0 (3.1.4) 2025-02-11 f4ab75096a OpenJDK 64-Bit Server VM 24.0.1+9-30 on 24.0.1+9-30 +jit [arm64-darwin]
Warming up --------------------------------------
                json    79.000 i/100ms
Calculating -------------------------------------
                json    814.479 (± 2.1%) i/s    (1.23 ms/i) -      8.216k in  10.092101s

== Encoding twitter.json (466906 bytes)
jruby 9.4.12.0 (3.1.4) 2025-02-11 f4ab75096a OpenJDK 64-Bit Server VM 24.0.1+9-30 on 24.0.1+9-30 +jit [arm64-darwin]
Warming up --------------------------------------
                json   136.000 i/100ms
Calculating -------------------------------------
                json      1.358k (± 1.0%) i/s  (736.45 μs/i) -     13.600k in  10.016731s

== Encoding ohai.json (20147 bytes)
jruby 9.4.12.0 (3.1.4) 2025-02-11 f4ab75096a OpenJDK 64-Bit Server VM 24.0.1+9-30 on 24.0.1+9-30 +jit [arm64-darwin]
Warming up --------------------------------------
                json     2.246k i/100ms
Calculating -------------------------------------
                json     21.987k (± 1.6%) i/s   (45.48 μs/i) -    220.108k in  10.013722s
```

### master (as of commit 37e689094a0152e5c3c774a0ec16f2ab70548336)
```
% ONLY=json ruby -I"lib" benchmark/encoder-realworld.rb 

== Encoding activitypub.json (52595 bytes)
jruby 9.4.12.0 (3.1.4) 2025-02-11 f4ab75096a OpenJDK 64-Bit Server VM 24.0.1+9-30 on 24.0.1+9-30 +jit [arm64-darwin]
Warming up --------------------------------------
                json   951.000 i/100ms
Calculating -------------------------------------
                json      9.517k (± 0.8%) i/s  (105.08 μs/i) -     96.051k in  10.093716s

== Encoding citm_catalog.json (500298 bytes)
jruby 9.4.12.0 (3.1.4) 2025-02-11 f4ab75096a OpenJDK 64-Bit Server VM 24.0.1+9-30 on 24.0.1+9-30 +jit [arm64-darwin]
Warming up --------------------------------------
                json    84.000 i/100ms
Calculating -------------------------------------
                json    843.486 (± 1.1%) i/s    (1.19 ms/i) -      8.484k in  10.059526s

== Encoding twitter.json (466906 bytes)
jruby 9.4.12.0 (3.1.4) 2025-02-11 f4ab75096a OpenJDK 64-Bit Server VM 24.0.1+9-30 on 24.0.1+9-30 +jit [arm64-darwin]
Warming up --------------------------------------
                json   145.000 i/100ms
Calculating -------------------------------------
                json      1.448k (± 0.8%) i/s  (690.73 μs/i) -     14.500k in  10.016276s

== Encoding ohai.json (20147 bytes)
jruby 9.4.12.0 (3.1.4) 2025-02-11 f4ab75096a OpenJDK 64-Bit Server VM 24.0.1+9-30 on 24.0.1+9-30 +jit [arm64-darwin]
Warming up --------------------------------------
                json     2.342k i/100ms
Calculating -------------------------------------
                json     23.073k (± 0.8%) i/s   (43.34 μs/i) -    231.858k in  10.049473s
```