@samyron 2025-08-14
Synthetic benchmarks of encoding an array of 128-byte ASCII strings.

```
benchmark_encoding "bytes.128.bestcase", ([("a" * 128)] * 10000)
```

### SegmetedByteListDirectOutputStream + SWAR
```
% ONLY=json JAVA_OPTS='-Djson.useSegmentedOutputStream=true -Djson.useSWARBasicEncoder=true' ruby -I"lib" benchmark/encoder-synthetic.rb

== Encoding bytes.128.bestcase (1310001 bytes)
jruby 9.4.12.0 (3.1.4) 2025-02-11 f4ab75096a OpenJDK 64-Bit Server VM 24.0.1+9-30 on 24.0.1+9-30 +jit [arm64-darwin]
Warming up --------------------------------------
                json   256.000 i/100ms
Calculating -------------------------------------
                json      2.561k (± 0.9%) i/s  (390.48 μs/i) -     25.600k in   9.997219s
```

### ByteListDirectOutputStream + Scalar (effectively the same code as master)
```
% ONLY=json JAVA_OPTS='-Djson.useSegmentedOutputStream=false -Djson.useSWARBasicEncoder=false' ruby -I"lib" benchmark/encoder-synthetic.rb

== Encoding bytes.128.bestcase (1310001 bytes)
jruby 9.4.12.0 (3.1.4) 2025-02-11 f4ab75096a OpenJDK 64-Bit Server VM 24.0.1+9-30 on 24.0.1+9-30 +jit [arm64-darwin]
Warming up --------------------------------------
                json   137.000 i/100ms
Calculating -------------------------------------
                json      1.376k (± 1.2%) i/s  (726.60 μs/i) -     13.837k in  10.055507s
```

### SegmentedByteListDirectOutputStream + Scalar
```
% ONLY=json JAVA_OPTS='-Djson.useSegmentedOutputStream=true -Djson.useSWARBasicEncoder=false' ruby -I"lib" benchmark/encoder-synthetic.rb

== Encoding bytes.128.bestcase (1310001 bytes)
jruby 9.4.12.0 (3.1.4) 2025-02-11 f4ab75096a OpenJDK 64-Bit Server VM 24.0.1+9-30 on 24.0.1+9-30 +jit [arm64-darwin]
Warming up --------------------------------------
                json   141.000 i/100ms
Calculating -------------------------------------
                json      1.424k (± 0.8%) i/s  (702.28 μs/i) -     14.241k in  10.001896s
```

### ByteListDirectOutputStream + SWAR

```
% ONLY=json JAVA_OPTS='-Djson.useSegmentedOutputStream=false -Djson.useSWARBasicEncoder=true' ruby -I"lib" benchmark/encoder-synthetic.rb

== Encoding bytes.128.bestcase (1310001 bytes)
jruby 9.4.12.0 (3.1.4) 2025-02-11 f4ab75096a OpenJDK 64-Bit Server VM 24.0.1+9-30 on 24.0.1+9-30 +jit [arm64-darwin]
Warming up --------------------------------------
                json   254.000 i/100ms
Calculating -------------------------------------
                json      2.558k (± 1.5%) i/s  (390.92 μs/i) -     25.654k in  10.030970s
```

### Master

```
% ONLY=json ruby -I"lib" benchmark/encoder-synthetic.rb

== Encoding bytes.128.bestcase (1310001 bytes)
jruby 9.4.12.0 (3.1.4) 2025-02-11 f4ab75096a OpenJDK 64-Bit Server VM 24.0.1+9-30 on 24.0.1+9-30 +jit [arm64-darwin]
Warming up --------------------------------------
                json   134.000 i/100ms
Calculating -------------------------------------
                json      1.334k (± 3.6%) i/s  (749.69 μs/i) -     13.400k in  10.062253s
```